requirements.txt
# ================================================================================
# REQUERIMIENTOS — SALUD MENTAL EN EL TRABAJO REMOTO
# ================================================================================
# Dependencias Python para el pipeline ETL
# Instalar con: pip install -r requirements.txt
# ================================================================================

# ── Core Data ────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── Base de Datos ────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# ── Configuración y Logging ───────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2

# ── Calidad de Datos ─────────────────────────────────────────────
great-expectations==0.18.14

# ── Pruebas ──────────────────────────────────────────────────────
pytest==8.2.1
pytest-cov==5.0.0

# ── Linting y Formatting ─────────────────────────────────────────
ruff==0.3.0

# ── Utilidades ───────────────────────────────────────────────────
tqdm==4.66.4
